requirements.txt
python-telegram-bot>=20.0
Pillow>=10.0
firebase-admin